support timeout in `Download`
The idea is to specify the maximum number of seconds to wait for
a download, via the optional component `maxTime` in the optional
second argument of `Download`.
The default value is given by the new user preference `DownloadMaxTime`
of the utils package, and the default for that is zero,
meaning that there is no timeout.

The methods stored in `Download_Methods` either use a given nonzero `maxTime`
value or signal failure (the latter holds for the method from the IO package,
which does not support timeouts).